+++ src/main/java/com/ditrit/letomodelizerapi/controller/ScopeController.java
@@ -45,17 +45,17 @@ import java.util.Map;
 import java.util.UUID;
 
 /**
- * REST Controller for managing groups.
- * Provides endpoints for CRUD operations on groups, including listing, retrieving, creating, updating, and deleting
- * groups.
+ * REST Controller for managing scope.
+ * Provides endpoints for CRUD operations on scopes, including listing, retrieving, creating, updating, and deleting
+ * scopes.
  * Only accessible by users with administrative permissions.
  */
 @Slf4j
-@Path("/groups")
+@Path("/scopes")
 @Produces(MediaType.APPLICATION_JSON)
 @Controller
 @AllArgsConstructor(onConstructor = @__(@Autowired))
-public class GroupController implements DefaultController {
+public class ScopeController implements DefaultController {
 
     /**
      * Service to manage user.
@@ -78,11 +78,11 @@ public class GroupController implements DefaultController {
     private AccessControlPermissionService accessControlPermissionService;
 
     /**
-     * Finds and returns all groups based on the provided query filters.
+     * Finds and returns all scopes based on the provided query filters.
      *
-     * @param request     HttpServletRequest to access the HTTP session
-     * @param uriInfo     UriInfo to extract query parameters
-     * @param queryFilter BeanParam for pagination and filtering
+     * @param request      HttpServletRequest to access the HTTP session
+     * @param uriInfo      UriInfo to extract query parameters
+     * @param queryFilter  BeanParam for pagination and filtering
      * @return a Response containing a page of AccessControlDTO objects
      */
     @GET
@@ -94,68 +94,68 @@ public class GroupController implements DefaultController {
         userPermissionService.checkIsAdmin(user, null);
         Map<String, String> filters = this.getFilters(uriInfo);
 
-        log.info("[{}] Received GET request to get groups with the following filters: {}", user.getLogin(), filters);
+        log.info("[{}] Received GET request to get scopes with the following filters: {}", user.getLogin(), filters);
         Page<AccessControlDTO> resources = accessControlService
-                .findAll(AccessControlType.GROUP, filters, queryFilter.getPagination())
+                .findAll(AccessControlType.SCOPE, filters, queryFilter.getPagination())
                 .map(new BeanMapper<>(AccessControlDTO.class));
 
         return Response.status(this.getStatus(resources)).entity(resources).build();
     }
 
     /**
-     * Retrieves a specific group by its ID.
+     * Retrieves a specific scope by its ID.
      *
      * @param request HttpServletRequest to access the HTTP session
-     * @param id      the ID of the group to retrieve
-     * @return a Response containing the AccessControlDTO of the requested group
+     * @param id      the ID of the scope to retrieve
+     * @return a Response containing the AccessControlDTO of the requested scope
      */
     @GET
     @Path("/{id}")
-    public Response getGroupById(final @Context HttpServletRequest request,
-                                 final @PathParam("id") @Valid @NotNull UUID id) {
+    public Response getScopeById(final @Context HttpServletRequest request,
+                                final @PathParam("id") @Valid @NotNull UUID id) {
         HttpSession session = request.getSession();
         User user = userService.getFromSession(session);
         userPermissionService.checkIsAdmin(user, null);
 
-        log.info("[{}] Received GET request to get group with id {}", user.getLogin(), id);
+        log.info("[{}] Received GET request to get scope with id {}", user.getLogin(), id);
         AccessControlDTO accessControlDTO = new BeanMapper<>(AccessControlDTO.class)
-                .apply(accessControlService.findById(AccessControlType.GROUP, id));
+                .apply(accessControlService.findById(AccessControlType.SCOPE, id));
 
         return Response.ok(accessControlDTO).build();
     }
 
     /**
-     * Creates a new group with the given details.
+     * Creates a new scope with the given details.
      *
-     * @param request             HttpServletRequest to access the HTTP session
-     * @param accessControlRecord Data for creating the new group
+     * @param request              HttpServletRequest to access the HTTP session
+     * @param accessControlRecord  Data for creating the new scope
      * @return a Response indicating the outcome of the creation operation
      */
     @POST
-    public Response createGroup(final @Context HttpServletRequest request,
-                                final @Valid AccessControlRecord accessControlRecord) {
+    public Response createScope(final @Context HttpServletRequest request,
+                               final @Valid AccessControlRecord accessControlRecord) {
         HttpSession session = request.getSession();
         User user = userService.getFromSession(session);
         userPermissionService.checkIsAdmin(user, null);
 
-        log.info("[{}] Received POST request to create a group with {}", user.getLogin(), accessControlRecord);
+        log.info("[{}] Received POST request to create a scope with {}", user.getLogin(), accessControlRecord);
         AccessControlDTO accessControlDTO = new BeanMapper<>(AccessControlDTO.class)
-                .apply(accessControlService.create(AccessControlType.GROUP, accessControlRecord));
+                .apply(accessControlService.create(AccessControlType.SCOPE, accessControlRecord));
 
         return Response.status(HttpStatus.CREATED.value()).entity(accessControlDTO).build();
     }
 
     /**
-     * Updates an existing group identified by the given ID with the provided details.
+     * Updates an existing scope identified by the given ID with the provided details.
      *
-     * @param request             HttpServletRequest to access the HTTP session
-     * @param id                  ID of the group to update
-     * @param accessControlRecord Data for updating the group
+     * @param request              HttpServletRequest to access the HTTP session
+     * @param id                   ID of the scope to update
+     * @param accessControlRecord  Data for updating the scope
      * @return a Response indicating the outcome of the update operation
      */
     @PUT
     @Path("/{id}")
-    public Response updateGroup(final @Context HttpServletRequest request,
+    public Response updateScope(final @Context HttpServletRequest request,
                                 final @PathParam("id") @Valid @NotNull UUID id,
                                 final @Valid AccessControlRecord accessControlRecord) {
         HttpSession session = request.getSession();
@@ -163,56 +163,56 @@ public class GroupController implements DefaultController {
         userPermissionService.checkIsAdmin(user, null);
 
         log.info(
-                "[{}] Received PUT request to update a group with id {} with {}",
+                "[{}] Received PUT request to update a scope with id {} with {}",
                 user.getLogin(),
                 id,
                 accessControlRecord
         );
 
         AccessControlDTO accessControlDTO = new BeanMapper<>(AccessControlDTO.class)
-                .apply(accessControlService.update(AccessControlType.GROUP, id, accessControlRecord));
+                .apply(accessControlService.update(AccessControlType.SCOPE, id, accessControlRecord));
 
         return Response.ok(accessControlDTO).build();
     }
 
     /**
-     * Deletes the group with the specified ID.
+     * Deletes the scope with the specified ID.
      *
      * @param request HttpServletRequest to access the HTTP session
-     * @param id      the ID of the group to delete
+     * @param id      the ID of the scope to delete
      * @return a Response indicating the outcome of the delete operation
      */
     @DELETE
     @Path("/{id}")
-    public Response deleteGroup(final @Context HttpServletRequest request,
+    public Response deleteScope(final @Context HttpServletRequest request,
                                 final @PathParam("id") @Valid @NotNull UUID id) {
         HttpSession session = request.getSession();
         User user = userService.getFromSession(session);
         userPermissionService.checkIsAdmin(user, null);
 
-        log.info("[{}] Received DELETE request to delete group with id {}", user.getLogin(), id);
-        accessControlService.delete(AccessControlType.GROUP, id);
+        log.info("[{}] Received DELETE request to delete scope with id {}", user.getLogin(), id);
+        accessControlService.delete(AccessControlType.SCOPE, id);
 
         return Response.noContent().build();
     }
 
     /**
-     * Retrieves users associated to a specific group.
-     * This endpoint fetches and returns a paginated list of users associated with the given group ID.
+     * Retrieves users associated to a specific scope.
+     * This endpoint fetches and returns a paginated list of users associated with the given scope ID.
      * The method is accessible only to users with administrative permissions.
      *
-     * @param request     HttpServletRequest to access the HTTP session for authentication and authorization.
-     * @param id          the ID of the group for which users are to be retrieved.
-     * @param uriInfo     UriInfo to extract query parameters for additional filtering.
-     * @param queryFilter BeanParam object for pagination and filtering purposes.
-     * @return a Response containing a paginated list of UserDTO objects associated with the group.
+     * @param request      HttpServletRequest to access the HTTP session for authentication and authorization.
+     * @param id           the ID of the scope for which users are to be retrieved.
+     * @param uriInfo      UriInfo to extract query parameters for additional filtering.
+     * @param queryFilter  BeanParam object for pagination and filtering purposes.
+     * @return a Response containing a paginated list of UserDTO objects associated with the scope.
      */
     @GET
     @Path("/{id}/users")
-    public Response getUsersByGroup(final @Context HttpServletRequest request,
-                                    final @PathParam("id") @Valid @NotNull UUID id,
-                                    final @Context UriInfo uriInfo,
-                                    final @BeanParam @Valid QueryFilter queryFilter) {
+    public Response getUsersByScope(final @Context HttpServletRequest request,
+                                   final @PathParam("id") @Valid @NotNull UUID id,
+                                   final @Context UriInfo uriInfo,
+                                   final @BeanParam @Valid QueryFilter queryFilter) {
         HttpSession session = request.getSession();
         User user = userService.getFromSession(session);
         userPermissionService.checkIsAdmin(user, null);
@@ -220,27 +220,27 @@ public class GroupController implements DefaultController {
         Map<String, String> filters = this.getFilters(uriInfo);
 
         log.info(
-                "[{}] Received GET request to get users of group {} with the following filters: {}",
+                "[{}] Received GET request to get users of scope {} with the following filters: {}",
                 user.getLogin(),
                 id,
                 filters
         );
 
         Page<UserDTO> resources = accessControlService
-                .findAllUsers(AccessControlType.GROUP, id, filters, queryFilter.getPagination())
+                .findAllUsers(AccessControlType.SCOPE, id, filters, queryFilter.getPagination())
                 .map(new BeanMapper<>(UserDTO.class));
 
         return Response.status(this.getStatus(resources)).entity(resources).build();
     }
 
     /**
-     * Associates a user to a specified group.
+     * Associates a user to a specified scope.
      * This endpoint allows an administrative user to associate another user, identified by their login,
-     * with a specific group, identified by its ID.
+     * with a specific scope, identified by its ID.
      *
      * @param request HttpServletRequest to access the HTTP session for authentication and authorization.
-     * @param id      the ID of the group to which the user is to be associated.
-     * @param login   the login identifier of the user to associate with the group.
+     * @param id      the ID of the scope to which the user is to be associated.
+     * @param login   the login identifier of the user to associate with the scope.
      * @return a Response indicating the outcome of the association operation.
      */
     @POST
@@ -253,53 +253,53 @@ public class GroupController implements DefaultController {
         User user = userService.getFromSession(session);
         userPermissionService.checkIsAdmin(user, null);
 
-        log.info("[{}] Received POST request to associate group {} with user {}", user.getLogin(), id, login);
-        accessControlService.associateUser(AccessControlType.GROUP, id, login);
+        log.info("[{}] Received POST request to associate scope {} with user {}", user.getLogin(), id, login);
+        accessControlService.associateUser(AccessControlType.SCOPE, id, login);
 
         return Response.status(HttpStatus.CREATED.value()).build();
     }
 
     /**
-     * Dissociates a user from a specified group.
+     * Dissociates a user from a specified scope.
      * This endpoint allows an administrative user to remove the association of a user, identified by their login,
-     * from a specific group, identified by its ID.
+     * from a specific scope, identified by its ID.
      *
      * @param request HttpServletRequest to access the HTTP session for authentication and authorization.
-     * @param id      the ID of the group from which the user is to be dissociated.
-     * @param login   the login identifier of the user to dissociate from the group.
+     * @param id      the ID of the scope from which the user is to be dissociated.
+     * @param login   the login identifier of the user to dissociate from the scope.
      * @return a Response indicating the outcome of the dissociation operation.
      */
     @DELETE
     @Path("/{id}/users/{login}")
     public Response dissociateUser(final @Context HttpServletRequest request,
-                                   final @PathParam("id") @Valid @NotNull UUID id,
-                                   final @PathParam(Constants.DEFAULT_USER_PROPERTY) @Valid @NotBlank String login) {
+                                  final @PathParam("id") @Valid @NotNull UUID id,
+                                  final @PathParam(Constants.DEFAULT_USER_PROPERTY) @Valid @NotBlank String login) {
         HttpSession session = request.getSession();
         User user = userService.getFromSession(session);
         userPermissionService.checkIsAdmin(user, null);
 
-        log.info("[{}] Received DELETE request to dissociate group {} with user {}", user.getLogin(), id, login);
-        accessControlService.dissociateUser(AccessControlType.GROUP, id, login);
+        log.info("[{}] Received DELETE request to dissociate scope {} with user {}", user.getLogin(), id, login);
+        accessControlService.dissociateUser(AccessControlType.SCOPE, id, login);
 
         return Response.noContent().build();
     }
 
     /**
-     * Retrieves the sub-groups of a specified group.
+     * Retrieves the groups of a specified scope.
      *
-     * <p>This method processes a GET request to obtain sub-groups associated with a given group ID. It filters the
-     * groups based on the provided query parameters and pagination settings.
+     * <p>This method processes a GET request to obtain groups associated with a given scope ID. It filters the
+     * scopes based on the provided query parameters and pagination settings.
      *
-     * @param request     the HttpServletRequest from which to obtain the HttpSession for user validation.
-     * @param id          the ID of the group for which to retrieve sub-groups. Must be a valid and non-null Long value.
-     * @param uriInfo     UriInfo context to extract query parameters for filtering results.
+     * @param request the HttpServletRequest from which to obtain the HttpSession for user validation.
+     * @param id the ID of the scope for which to retrieve sub-scopes. Must be a valid and non-null Long value.
+     * @param uriInfo UriInfo context to extract query parameters for filtering results.
      * @param queryFilter bean parameter encapsulating filtering and pagination criteria.
      * @return a Response object containing the requested page of AccessControlDirectDTO objects representing the
-     * sub-groups of the specified group. The status of the response can vary based on the outcome of the request.
+     * groups of the specified scope. The status of the response can vary based on the outcome of the request.
      */
     @GET
     @Path("/{id}/groups")
-    public Response getSubGroupsOfGroup(final @Context HttpServletRequest request,
+    public Response getGroupsOfScope(final @Context HttpServletRequest request,
                                         final @PathParam("id") @Valid @NotNull UUID id,
                                         final @Context UriInfo uriInfo,
                                         final @BeanParam @Valid QueryFilter queryFilter) {
@@ -310,7 +310,7 @@ public class GroupController implements DefaultController {
         Map<String, String> filters = this.getFilters(uriInfo);
 
         log.info(
-                "[{}] Received GET request to get sub-groups of group {} with the following filters: {}",
+                "[{}] Received GET request to get groups of scope {} with the following filters: {}",
                 user.getLogin(),
                 id,
                 filters
@@ -318,7 +318,7 @@ public class GroupController implements DefaultController {
 
         Page<AccessControlDirectDTO> resources = accessControlService
                 .findAllAccessControls(
-                        AccessControlType.GROUP,
+                        AccessControlType.SCOPE,
                         id,
                         AccessControlType.GROUP,
                         filters,
@@ -329,17 +329,15 @@ public class GroupController implements DefaultController {
     }
 
     /**
-     * Associates a group with another group.
+     * Associates a group with a scope.
      *
      * <p>This method handles a POST request to create an association between two groups, specified by their IDs.
      * It validates the user's session and ensures the user has administrative privileges before proceeding with the
      * association.
      *
-     * @param request    the HttpServletRequest from which to obtain the HttpSession for user validation.
-     * @param id         the ID of the first group to which the second group will be associated. Must be a valid and
-     *                   non-null Long value.
-     * @param subGroupId the ID of the second group to be associated with the first group. Must be a valid and non-null
-     *                   Long value.
+     * @param request the HttpServletRequest from which to obtain the HttpSession for user validation.
+     * @param id the ID of the scope to which the group will be associated. Must be a valid and non-nul Long value.
+     * @param groupId the ID of the group to be associated with the scope. Must be a valid and non-null Long value.
      * @return a Response object indicating the outcome of the association operation. A successful operation returns
      * a status of CREATED.
      */
@@ -348,79 +346,78 @@ public class GroupController implements DefaultController {
     @Path("/{id}/groups")
     public Response associate(final @Context HttpServletRequest request,
                               final @PathParam("id") @Valid @NotNull UUID id,
-                              final @Valid @NotNull String subGroupId) {
+                              final @Valid @NotNull String groupId) {
         HttpSession session = request.getSession();
         User user = userService.getFromSession(session);
         userPermissionService.checkIsAdmin(user, null);
 
-        log.info("[{}] Received POST request to associate group {} with group {}", user.getLogin(), id, subGroupId);
+        log.info("[{}] Received POST request to associate scope {} with group {}", user.getLogin(), id, groupId);
         accessControlService.associate(
-                AccessControlType.GROUP,
+                AccessControlType.SCOPE,
                 id,
                 AccessControlType.GROUP,
-                UUID.fromString(subGroupId)
+                UUID.fromString(groupId)
         );
 
         return Response.status(HttpStatus.CREATED.value()).build();
     }
 
     /**
-     * Dissociates a group from another group.
+     * Dissociates a group from a scope.
      *
-     * <p>This method facilitates the handling of a DELETE request to remove the association between two groups,
+     * <p>This method facilitates the handling of a DELETE request to remove the association between group and scope,
      * identified by their respective IDs. The operation is secured, requiring validation of the user's session and
      * administrative privileges.
      *
-     * @param request    the HttpServletRequest used to validate the user's session.
-     * @param id         the ID of the first group from which the second group will be dissociated. Must be a valid and
-     *                   non-null Long value.
-     * @param subGroupId the ID of the second group to be dissociated from the first group. Must be a valid and non-null
-     *                   Long value.
+     * @param request the HttpServletRequest used to validate the user's session.
+     * @param id the ID of the scope from which the group will be dissociated. Must be a valid and non-null Long value.
+     * @param groupId the ID of the second scope to be dissociated from the first scope. Must be a valid and non-null
+     *                Long value.
      * @return a Response object with a status indicating the outcome of the dissociation operation. A successful
      * operation returns a status of NO_CONTENT.
      */
     @DELETE
-    @Path("/{id}/groups/{subGroupId}")
+    @Path("/{id}/groups/{groupId}")
     public Response dissociate(final @Context HttpServletRequest request,
                                final @PathParam("id") @Valid @NotNull UUID id,
-                               final @PathParam("subGroupId") @Valid @NotNull UUID subGroupId) {
+                               final @PathParam("groupId") @Valid @NotNull UUID groupId) {
         HttpSession session = request.getSession();
         User user = userService.getFromSession(session);
         userPermissionService.checkIsAdmin(user, null);
 
         log.info(
-                "[{}] Received DELETE request to dissociate sub-group {} with group {}",
+                "[{}] Received DELETE request to dissociate scope {} with group {}",
                 user.getLogin(),
                 id,
-                subGroupId
+                groupId
         );
 
-        accessControlService.dissociate(AccessControlType.GROUP, id, AccessControlType.GROUP, subGroupId);
+        accessControlService.dissociate(AccessControlType.SCOPE, id, AccessControlType.SCOPE, groupId);
 
         return Response.noContent().build();
     }
 
     /**
-     * Retrieves the roles associated with a specific group, identified by the group's ID. This method processes a GET
+     * Retrieves the roles associated with a specific scope, identified by the scope's ID. This method processes a GET
      * request and returns a paginated list of AccessControlDirectDTO objects representing the roles. It supports
-     * filtering based on query parameters to allow for refined searches within the group's associated roles.
+     * filtering based on query parameters to allow for refined searches within the scope's associated roles.
      *
      * <p>This endpoint is secured to ensure that only users with administrative permissions can access role information
-     * for a group. It is particularly useful for managing and reviewing the roles and permissions assigned to a group
+     * for a scope. It is particularly useful for managing and reviewing the roles and permissions assigned to a scope
      * within an access control system.
      *
-     * @param request     The HttpServletRequest providing context to access the HTTP session.
-     * @param id          The unique identifier of the group whose roles are being retrieved.
-     * @param uriInfo     URI information for extracting query parameters to apply as filters.
+     * @param request The HttpServletRequest providing context to access the HTTP session.
+     * @param id The unique identifier of the scope whose roles are being retrieved.
+     * @param uriInfo URI information for extracting query parameters to apply as filters.
      * @param queryFilter A BeanParam object encapsulating pagination and filtering criteria to manage result set size
      *                    and relevance.
      * @return A Response object containing a paginated list of AccessControlDirectDTO objects representing the roles
-     * associated with the specified group. The response includes pagination details and adheres to the
-     * HTTP status code conventions to indicate the outcome of the request.
+     *         associated with the specified scope. The response includes pagination details and adheres to the
+     *         HTTP status code conventions to indicate the outcome of the request.
      */
     @GET
     @Path("/{id}/roles")
-    public Response getRolesOfGroup(final @Context HttpServletRequest request,
+    public Response getRolesOfScope(final @Context HttpServletRequest request,
                                     final @PathParam("id") @Valid @NotNull UUID id,
                                     final @Context UriInfo uriInfo,
                                     final @BeanParam @Valid QueryFilter queryFilter) {
@@ -431,7 +428,7 @@ public class GroupController implements DefaultController {
         Map<String, String> filters = this.getFilters(uriInfo);
 
         log.info(
-                "[{}] Received GET request to get roles of group {} with the following filters: {}",
+                "[{}] Received GET request to get roles of scope {} with the following filters: {}",
                 user.getLogin(),
                 id,
                 filters
@@ -439,57 +436,9 @@ public class GroupController implements DefaultController {
 
         Page<AccessControlDirectDTO> resources = accessControlService
                 .findAllAccessControls(
-                        AccessControlType.GROUP,
-                        id, AccessControlType.ROLE,
-                        filters,
-                        queryFilter.getPagination()
-                );
-
-        return Response.status(this.getStatus(resources)).entity(resources).build();
-    }
-
-    /**
-     * Retrieves the scopes associated with a specific group, identified by the group's ID. This method processes a GET
-     * request and returns a paginated list of AccessControlDirectDTO objects representing the scopes. It supports
-     * filtering based on query parameters to allow for refined searches within the group's associated scopes.
-     *
-     * <p>This endpoint is secured to ensure that only users with administrative permissions can access scope
-     * information for a group. It is particularly useful for managing and reviewing the roles and permissions assigned
-     * to a group within an access control system.
-     *
-     * @param request     The HttpServletRequest providing context to access the HTTP session.
-     * @param id          The unique identifier of the group whose scopes are being retrieved.
-     * @param uriInfo     URI information for extracting query parameters to apply as filters.
-     * @param queryFilter A BeanParam object encapsulating pagination and filtering criteria to manage result set size
-     *                    and relevance.
-     * @return A Response object containing a paginated list of AccessControlDirectDTO objects representing the scopes
-     * associated with the specified group. The response includes pagination details and adheres to the
-     * HTTP status code conventions to indicate the outcome of the request.
-     */
-    @GET
-    @Path("/{id}/scopes")
-    public Response getScopesOfGroup(final @Context HttpServletRequest request,
-                                    final @PathParam("id") @Valid @NotNull UUID id,
-                                    final @Context UriInfo uriInfo,
-                                    final @BeanParam @Valid QueryFilter queryFilter) {
-        HttpSession session = request.getSession();
-        User user = userService.getFromSession(session);
-        userPermissionService.checkIsAdmin(user, null);
-
-        Map<String, String> filters = this.getFilters(uriInfo);
-
-        log.info(
-                "[{}] Received GET request to get scopes of group {} with the following filters: {}",
-                user.getLogin(),
-                id,
-                filters
-        );
-
-        Page<AccessControlDirectDTO> resources = accessControlService
-                .findAllAccessControls(
-                        AccessControlType.GROUP,
-                        id,
                         AccessControlType.SCOPE,
+                        id,
+                        AccessControlType.ROLE,
                         filters,
                         queryFilter.getPagination()
                 );
@@ -498,22 +447,21 @@ public class GroupController implements DefaultController {
     }
 
     /**
-     * Retrieves the permissions of a specified group.
+     * Retrieves the permissions of a specified scope.
      *
-     * <p>This method processes a GET request to obtain permissions associated with a given group ID. It filters the
+     * <p>This method processes a GET request to obtain permissions associated with a given scope ID. It filters the
      * permissions based on the provided query parameters and pagination settings.
      *
-     * @param request     the HttpServletRequest from which to obtain the HttpSession for user validation.
-     * @param id          the ID of the groups for which to retrieve permissions. Must be a valid and non-null Long
-     *                    value.
-     * @param uriInfo     UriInfo context to extract query parameters for filtering results.
+     * @param request the HttpServletRequest from which to obtain the HttpSession for user validation.
+     * @param id the ID of the scopes for which to retrieve permissions. Must be a valid and non-null Long value.
+     * @param uriInfo UriInfo context to extract query parameters for filtering results.
      * @param queryFilter bean parameter encapsulating filtering and pagination criteria.
      * @return a Response object containing the requested page of PermissionDirectDTO objects representing the
-     * permissions of the specified group. The status of the response can vary based on the outcome of the request.
+     * permissions of the specified scope. The status of the response can vary based on the outcome of the request.
      */
     @GET
     @Path("/{id}/permissions")
-    public Response getPermissionsOfGroup(final @Context HttpServletRequest request,
+    public Response getPermissionsOfScope(final @Context HttpServletRequest request,
                                           final @PathParam("id") @Valid @NotNull UUID id,
                                           final @Context UriInfo uriInfo,
                                           final @BeanParam @Valid QueryFilter queryFilter) {
@@ -524,13 +472,13 @@ public class GroupController implements DefaultController {
         Map<String, String> filters = this.getFilters(uriInfo);
 
         log.info(
-                "[{}] Received GET request to get permissions of group {} with the following filters: {}",
+                "[{}] Received GET request to get permissions of scope {} with the following filters: {}",
                 user.getLogin(),
                 id,
                 filters
         );
 
-        AccessControl accessControl = accessControlService.findById(AccessControlType.GROUP, id);
+        AccessControl accessControl = accessControlService.findById(AccessControlType.SCOPE, id);
         Page<PermissionDirectDTO> resources = accessControlPermissionService
                 .findAll(accessControl.getId(), filters, queryFilter.getPagination())
                 .map(new AccessControlPermissionViewToPermissionDirectDTOFunction());